x64 codegen: consume RaProgram directly + text test infra (#119, #121)
Split x64 and aarch64 backends into separate files. The x64 backend now
iterates RaProgram blocks directly instead of the flat LinearIr op list,
eliminating the diamond dependency where codegen needed both LinearIr and
AllocatedProgram.

Add public APIs for compiling from IR text and RA-MIR text, enabling the
bug minimization workflow: paste a failing snapshot, strip it down to a
minimal reproducer, re-run.

RA-MIR parser: add /hw{n} fixed-reg support, assign contiguous
linear_op_index across blocks for correct regalloc2 edge moves.
